added 7th

+++ quick_sort.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 
-void swap(int* a, int* b) {
-    int t = *a;
+void swap(double* a, double* b) {
+    double t = *a;
     *a = *b;
     *b = t;
 }
 
-int partition(int arr[], int low, int high) {
-    int pivot = arr[high];
+int partition(double arr[], int low, int high) {
+    double pivot = arr[high];
     int i = (low - 1);
 
     for (int j = low; j <= high - 1; j++) {
@@ -20,14 +20,14 @@ int partition(int arr[], int low, int high) {
     return (i + 1);
 }
 
-void quicksort(int arr[], int low, int high) {
+void quicksort(double arr[], int low, int high) {
     if (low < high) {
         int pi = partition(arr, low, high);
         quicksort(arr, low, pi - 1);
         quicksort(arr, pi + 1, high);
     }
 }
-
+/*
 int main() {
     int arr[] = { 10, 7, 8, 9, 1, 5 };
     int n = sizeof(arr) / sizeof(arr[0]);
@@ -41,3 +41,4 @@ int main() {
     printf("\n");
     return 0;
 }
+*/
